Note
This is my first try for IOS app. i have not took any class on programing before but i really wanted to be do something new on programming.
i feel like code path brought a super opportunity in front of me. May be my program is not good at all but juts to figure our this all it took me more then 40 hour on doing research on internet and friend.  i would really grateful if i can get in to a code path boot came and get an opportunity to build an app for IOS device. hope to see again soon and learn something new to make something new.